Put item availability together with instance info/links
